update namespacing for facilities/metadata client & configuration
